fix(FE): [KC-17] implement metadata system and component architecture
