+++ .gitignore
@@ -43,3 +43,5 @@ next-env.d.ts
 # todo
 todo
 
+
+certificates
