+++ src/app/auth/login/page.tsx
@@ -1,13 +1,12 @@
-import { Suspense } from 'react';
+import type { Metadata } from 'next';
 
 import { LoginForm } from '@/components/auth/login-form';
 
-const LoginPage = () => {
-  return (
-    <Suspense fallback={<div>Loading...</div>}>
-      <LoginForm />
-    </Suspense>
-  );
+export const metadata: Metadata = {
+  title: 'Login',
+  description: 'Login to your Kurl Club Admin account to manage your gym',
 };
 
-export default LoginPage;
+export default function LoginPage() {
+  return <LoginForm />;
+}
