+++ src/app/auth/register/page.tsx
@@ -1,4 +1,11 @@
+import type { Metadata } from 'next';
+
 import { RegisterForm } from '@/components/auth/register-form';
+
+export const metadata: Metadata = {
+  title: 'Register',
+  description: 'Create your Kurl Club Admin account to start managing your gym',
+};
 
 const RegisterPage = () => {
   return <RegisterForm />;
