+++ src/app/auth/reset/page.tsx
@@ -1,4 +1,11 @@
+import type { Metadata } from 'next';
+
 import { ResetForm } from '@/components/auth/reset-form';
+
+export const metadata: Metadata = {
+  title: 'Reset Password',
+  description: 'Reset your Kurl Club Admin account password',
+};
 
 const ResetPage = () => {
   return <ResetForm />;
