+++ src/app/payments-and-attendance/(tabs)/attendance/page.tsx
@@ -1,3 +1,10 @@
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'Attendance',
+  description: 'Track member attendance and gym visit records',
+};
+
 const AttendancePage = () => {
   return <div>AttendancePage</div>;
 };
