+++ src/app/settings/(tabs)/membership-plans/page.tsx
@@ -1,4 +1,11 @@
+import type { Metadata } from 'next';
+
 import { PackageManager } from '@/components/settings/membership-planner';
+
+export const metadata: Metadata = {
+  title: 'Membership Plans',
+  description: 'Configure gym membership plans, pricing, and packages',
+};
 
 export default function MembershipPlanPage() {
   return (
