+++ src/components/payments/tabs/history.tsx
@@ -1,4 +1,4 @@
-import { PaymentsTab } from './payments-tab';
+import { PaymentsTab } from '../payments-tab';
 
 const History = () => {
   return <PaymentsTab type="history" />;
